pyproject.toml
[tool.poetry]
name = "zpipz"
version = "0.0.0"
description = "AWS CDK L3 Constructs"
authors = ["zpprado"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8"
aws-cdk-lib = ">=2.30.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-cov = "^3.0.0"

[tool.semantic_release]
version_variable = "pyproject.toml:version" # version location
branch = "main"                             # branch to make releases of
changelog_file = "CHANGELOG.md"             # changelog file
build_command = "poetry build"              # build dists
dist_path = "dist/"                         # where to put dists
upload_to_release = true                    # auto-create GitHub release
upload_to_pypi = false                      # don't auto-upload to PyPI
remove_dist = false                         # don't remove dists
patch_without_tag = true                    # patch release by default

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
